requirements.txt
paho-mqtt==2.1.0
influxdb-client==1.45.0
pandas==2.2.2
matplotlib==3.9.0
python-dotenv==1.0.1
